requirements.txt
langchain
python-dotenv
langchain_community
langchain-ollama
langchain-core
fastapi
uvicorn
langserve
sse_starlette